web/middleware/requirements.txt
celery==4.0.2
celery-once==1.2.0
Django==1.11.14
django-bootstrap-static==4.0.0
django-celery-beat==1.1.1
django-celery-monitor==1.1.2
django_celery_results==1.0.1
django-extensions==2.1.0
django-redis==4.9.0
django-rest-swagger==2.2.0
djangorestframework==3.8.2
django-cors-headers==2.4.0
gunicorn==19.9.0
psycopg2==2.7.5
redis==2.10.6
requests==2.19.1
coverage==4.5.1
kombu==4.0.2
billiard==3.5.0.2
boto==2.49.0
watchtower==0.5.3
boto3==1.8.6